Splitting things out and adding edit command
